pyproject.toml
[project]
name = "aladdin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "loguru>=0.7.3",
    "lxml>=6.0.1",
    "openai>=1.100.1",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pycryptodome>=3.23.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.6",
    "reflex-ag-grid>=0.0.11",
    "reflex-chakra>=0.8.2.post1",
    "reflex-chat>=0.0.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
